Fix tests
Tests left handles open. Timer in each test was nowhere near
enough as several seconds were needed. A proper fix was needed.
The timer hack was on github issue of supertest.

By default express opens a client range port which is closed when
calling listen. However, if supertest makes a request to the
initial socket before a second socket is bound there is no way
to control closing the initial socket.

In addition also the bound socket needs to be closed.

Logout test was also commented out. The test worked exactly as
written. Logout was not allowed because there was no logged in
user. Although 404 is clearly wrong status let's leave fixing it
for later.