Refactor backend to core, data & web directories
eslint rule is used to enforce import sanity. Core must not use data
or web and data must not use web. Logic should contain unit tested
business logic whereas data and web handle database and web
application matters respectively.

There are probably validation tests on web level that should be
moved to core along with comprehensive validation unit tests. One
such test is good on web level to ensure validation is included.
This change was left out of scope as the situation is not any worse
than it used to be.