Wrap react-redux
react-redux is problematic because
reduxjs/toolkit provides functionality that
is preferred over direct react-redux usage.
However, it is lacking Provider and useDispatch
hook and createSelector is not that useful. So
it's not possible to completely ditch
react-redux and that way ensure all the legacy
stuff such as connect stays out of the
application.