Upgrade @koa/router
In @koa/router 15.1.0 exported router was changed to refer to constructor
instead of the class. This makes Router more or less unusable as a type.

By moving router configuration to the function that creates it we don't
need to expose the type. This is better also in a way that @koa/router
dependency is referred to only in one file. Having to use koa in two files
is less of a problem as @koa/router depends on koa anyway whether
explicitly mentioned in code or not.